+++ ChangeLog.md
@@ -26,6 +26,7 @@
 - BUGFIX: #5395 Fixed SQL From File executor to handle properly temporar FK disabling.
 - BUGFIX: #5395 Fixed loading trigger condition if it starts with the NOT keyword.
 - BUGFIX: #5398 Fixed Smart Executor for queries, where two joined, unaliased tables return column of the same name, but different case-wise.
+- BUGFIX: #5406 Tracking calls to load_extension() SQL function and loading the same extensions in the Results Counting connection.
 - BUGFIX: Fixed arbitrary crashes when opening SQL Editor, while having non-thread-safe SQLite extensions loaded in the active database.
 
 ### 3.4.17